==> Microbiome_Analysis/5_stats/00_wrapper.Rmd <==
---
title: "Herbicide Effects - Microbiome-Analysis"
author: "Jan Waelchli"
date: '`r Sys.Date()`'
geometry: margin=2cm
output:
  pdf_document:
    toc: yes
    toc_depth: 2
---

```{r setup, include=FALSE}

#This file is used to knitr all chapter togehter as one pdf!

knitr::opts_chunk$set(echo = TRUE, fig.align="center", fig.path='figures_auto/')
options(knitr.duplicate.label = "allow", tinytex.verbose = TRUE)
```

```{r, child=c('01_Import_Normalization.Rmd', '02_Taxa_Analysis.Rmd', '03_Alpha_Diversity.Rmd', '04_Beta_Diversity.Rmd', '05_Taxa_Response.Rmd')}

#By knitting this script you receive one pdf out of all child Rmd-documents.

```
