==> Microbiome_Analysis/5_stats/02_Taxa_Analysis.Rmd <==
---
title: "Herbicide - 02: Taxanomy Analysis"
author: "Jan Waelchli"
geometry: margin=2cm
output:
  pdf_document:
    toc: yes
    toc_depth: 3
---


```{r setup, include=FALSE, echo=F, warning=F, message=F,}

##clear the object from memory
rm(list=ls())

#knitr settings
knitr::opts_chunk$set(echo=TRUE, fig.align="center")
options(tinytex.verbose = TRUE)

#set seed
set.seed(100)

## set source to file location
library(rstudioapi)
setwd(dirname(getActiveDocumentContext()$path))

## load functions
source("functions/functions.R")
functions(path="functions/")

## install (if necessary) and load libraries
libraries()

```

```{r import, echo=F, message=F, warning=F}

# Import

# set wd to RDS files
setwd("interim/01_Import_Normalization")

#cImport RDS files
bDESIGN <- readRDS("bDESIGN.RDS")
bTAX <- readRDS("bTAX.RDS")
bDAT_rare <- readRDS("bDAT_rare.RDS")
fDESIGN <- readRDS("fDESIGN.RDS")
fTAX <- readRDS("fTAX.RDS")
fDAT_rare <- readRDS("fDAT_rare.RDS")

bTAX_level_cols_phyla <- readRDS("bTAX_level_cols_phyla.RDS")
fTAX_level_cols_phyla <- readRDS("fTAX_level_cols_phyla.RDS")

```

```{r bac phyla, echo=F, warning=F, message=F}

### BACTERIA ###

# Data as bPHYLOSEQ object

#change order for melting later
bTAX <- bTAX[,c("labels", "kingdom", "phylum", "class", "order", "family", "genus", "ASV_ID", "cols_phyla")]

#create phyloseq objects
bDESIGN <- as.data.frame(bDESIGN)
rownames(bDESIGN) <- bDESIGN$Label
bDAT_rare <- as.matrix(bDAT_rare)
bTAX <- as.matrix(bTAX[rownames(bDAT_rare),])

bPHYSEQ <- phyloseq::phyloseq(sample_data(bDESIGN), 
                       otu_table(bDAT_rare, taxa_are_rows=T),
                       tax_table(bTAX))

# agglomerate data by labels 
bPHYSEQ_phyla <- tax_glom(bPHYSEQ, "labels") # merge by 'labels' 
bPHYSEQ_phyla <- transform_sample_counts(bPHYSEQ_phyla, function(x) 100 * x/sum(x)) 

# melt phyloseq object
bPHYSEQ_phyla_melt <- psmelt(bPHYSEQ_phyla) 

# Defining ASV colors by labels
bPHYSEQ_phyla_melt$labels <- as.factor(bPHYSEQ_phyla_melt$labels)
bPHYSEQ_phyla_melt$cols <- as.character(bPHYSEQ_phyla_melt$labels)

# attributing previously assigned colors
# bTAX_level_cols_phyla
for(i in names(bTAX_level_cols_phyla)[names(bTAX_level_cols_phyla) %in% levels(bPHYSEQ_phyla_melt$labels)]){
  bPHYSEQ_phyla_melt[bPHYSEQ_phyla_melt$labels==paste(i), ]$cols <- bTAX_level_cols_phyla[paste(i)]
}

# Defining high abundant Phyla
# Phyla with MEAN abundances higher than 1% relative abundances
bPHYSEQ_phyla_abu <-  rownames(otu_table(bPHYSEQ_phyla))[apply(otu_table(bPHYSEQ_phyla), 1, mean, na.rm=T) > 1]
bPHYSEQ_phyla_abuP <- tax_table(bPHYSEQ_phyla)[rownames(tax_table(bPHYSEQ_phyla)) %in% bPHYSEQ_phyla_abu, "labels"]

# Defining low abundant Phyla
# Phyla with MEAN abundances lower than 1% relative abundances
bPHYSEQ_phyla_low <-  rownames(otu_table(bPHYSEQ_phyla))[apply(otu_table(bPHYSEQ_phyla), 1, mean, na.rm=T) < 1]
bPHYSEQ_phyla_lowP <- tax_table(bPHYSEQ_phyla)[rownames(tax_table(bPHYSEQ_phyla)) %in% bPHYSEQ_phyla_low, "labels"]

# subsetting the color vector to abundant labels and classes
# delete labels name of low-abundant phyla (for plot) and put them at the bottom
bPHYSEQ_phyla_melt$labels_2 <- as.character(bPHYSEQ_phyla_melt$labels)
bPHYSEQ_phyla_melt[bPHYSEQ_phyla_melt$labels_2 %in% bPHYSEQ_phyla_lowP, ]$labels_2 <- "Low abundant phyla"
bPHYSEQ_phyla_melt$labels_2 <- as.factor(bPHYSEQ_phyla_melt$labels_2)
bPHYSEQ_phyla_melt$labels_2 <- factor(bPHYSEQ_phyla_melt$labels_2)

# color matrix for plot
bcol_class <- bPHYSEQ_phyla_melt$cols
names(bcol_class) <- bPHYSEQ_phyla_melt$labels_2

```

```{r fun phyla, echo=F, warning=F, message=F}

### Fungi ###

# Data as PHYLOSEQ object

#change order for melting later
fTAX <- fTAX[,c("labels", "kingdom", "phylum", "class", "order", "family", "genus", "ASV_ID", "cols_phyla")]

#create phyloseq objects
fDESIGN <- as.data.frame(fDESIGN)
rownames(fDESIGN) <- fDESIGN$Label
fDAT_rare <- as.matrix(fDAT_rare)
fTAX <- as.matrix(fTAX[rownames(fDAT_rare),])

fPHYSEQ <- phyloseq(sample_data(fDESIGN), 
                   otu_table(fDAT_rare, taxa_are_rows=T),
                   tax_table(fTAX))

# agglomerate data by labels 
fPHYSEQ_phyla <- tax_glom(fPHYSEQ, "labels") # merge by 'labels' 
fPHYSEQ_phyla <- transform_sample_counts(fPHYSEQ_phyla, function(x) 100 * x/sum(x)) 

# melt phyloseq object
fPHYSEQ_phyla_melt <- psmelt(fPHYSEQ_phyla) 

# Defining ASV colors by labels
fPHYSEQ_phyla_melt$labels <- as.factor(fPHYSEQ_phyla_melt$labels)
fPHYSEQ_phyla_melt$cols <- as.character(fPHYSEQ_phyla_melt$labels)

# attributing previously assigned colors
# fTAX_level_cols_phyla
for(i in names(fTAX_level_cols_phyla)[names(fTAX_level_cols_phyla) %in% levels(fPHYSEQ_phyla_melt$labels)]){
  fPHYSEQ_phyla_melt[fPHYSEQ_phyla_melt$labels==paste(i), ]$cols <- fTAX_level_cols_phyla[paste(i)]
}

# Defining high abundant Phyla
# Phyla with MEAN abundances higher than 1% relative abundances
fPHYSEQ_phyla_abu <-  rownames(otu_table(fPHYSEQ_phyla))[apply(otu_table(fPHYSEQ_phyla), 1, mean, na.rm=T) > 1]
fPHYSEQ_phyla_abuP <- tax_table(fPHYSEQ_phyla)[rownames(tax_table(fPHYSEQ_phyla)) %in% fPHYSEQ_phyla_abu, "labels"]

# Defining low abundant Phyla
# Phyla with MEAN abundances lower than 1% relative abundances
fPHYSEQ_phyla_low <-  rownames(otu_table(fPHYSEQ_phyla))[apply(otu_table(fPHYSEQ_phyla), 1, mean, na.rm=T) < 1]
fPHYSEQ_phyla_lowP <- tax_table(fPHYSEQ_phyla)[rownames(tax_table(fPHYSEQ_phyla)) %in% fPHYSEQ_phyla_low, "labels"]

# subsetting the color vector to abundant labels and classes
# delete labels name of low-abundant phyla (for plot) and put them at the bottom
fPHYSEQ_phyla_melt$labels_2 <- as.character(fPHYSEQ_phyla_melt$labels)
fPHYSEQ_phyla_melt[fPHYSEQ_phyla_melt$labels_2 %in% fPHYSEQ_phyla_lowP, ]$labels_2 <- "Low abundant phyla"
fPHYSEQ_phyla_melt$labels_2 <- as.factor(fPHYSEQ_phyla_melt$labels_2)
fPHYSEQ_phyla_melt$labels_2 <- factor(fPHYSEQ_phyla_melt$labels_2)

# color matrix for plot
fcol_class <- fPHYSEQ_phyla_melt$cols
names(fcol_class) <- fPHYSEQ_phyla_melt$labels_2

```

# Taxonomy

## Phyla abundance plot

We got an overview over the abundance of bacterial and fungal taxonomy by showing the most abundant phyla for each sample.

### Figure 2.1 | Bacteria: Phylum level taxonomy

\vspace{5mm}

```{r bac fig, echo=F, warning=F, message=F, fig.height=13, fig.width=15}


#sort
bPHYSEQ_phyla_melt$group <- factor(bPHYSEQ_phyla_melt$group, levels=c("ctr_soil","ctr_weeds","gly_soil","gly_weeds","tb_soil","tb_weeds"))
bPHYSEQ_phyla_melt$labels_2 <- forcats::fct_relevel(bPHYSEQ_phyla_melt$labels_2, "Low abundant phyla", after = 4)
bPHYSEQ_phyla_melt <- bPHYSEQ_phyla_melt[order(bPHYSEQ_phyla_melt$group),]
bPHYSEQ_phyla_melt <- bPHYSEQ_phyla_melt[order(bPHYSEQ_phyla_melt$Time),]
bPHYSEQ_phyla_melt <- bPHYSEQ_phyla_melt[order(bPHYSEQ_phyla_melt$Experiment),]

bPHYSEQ_phyla_melt$SampleID2 <- gsub("_(2|5)W_", "", bPHYSEQ_phyla_melt$SampleID)
levels <- unique(bPHYSEQ_phyla_melt$SampleID2)
bPHYSEQ_phyla_melt$SampleID2 <- factor(bPHYSEQ_phyla_melt$SampleID2, levels=levels)
bPHYSEQ_phyla_melt$Time <- as.factor(bPHYSEQ_phyla_melt$Time)
levels(bPHYSEQ_phyla_melt$Time) <- c("2W", "5W")
levels(bPHYSEQ_phyla_melt$Experiment) <- c("Experiment 1", "Experiment 2")

#plot
fig2.1 <- ggplot(bPHYSEQ_phyla_melt, aes_string(x="SampleID2", y="Abundance", fill="labels_2")) + 
  geom_bar(stat="identity")+
  xlab("") + 
  ylab("Relative abundance [%]") +
  scale_colour_manual(values=bcol_class)+
  scale_fill_manual(values=bcol_class) +
  guides(fill=guide_legend(title="Phylum"))+
  facet_grid(Time ~ Experiment, scale="free_x", space="free_x") +
  geom_text(data=bPHYSEQ_phyla_melt, aes(label=paste0(Herbicide, " ", Application), y=-15), angle=90, check_overlap=T, size=3, hjust = 0.25) + 
  theme_bw()+
  ggtitle("Bacteria: Taxonomy")+
  theme(axis.text.x=element_blank(), axis.ticks.x=element_blank())

#png("figures/fig2.1.png",10000,4500, res=600)
print(fig2.1)
#dev.off()

```
\vspace{5mm}

### Figure 2.2 | Fungi: Phylum level taxonomy


```{r fun fig, echo=F, warning=F, message=F, fig.height=13, fig.width=15}

#sort
fPHYSEQ_phyla_melt$group <- factor(fPHYSEQ_phyla_melt$group, levels=c("ctr_soil","ctr_weeds","gly_soil","gly_weeds","tb_soil","tb_weeds"))
fPHYSEQ_phyla_melt$labels_2 <- fct_relevel(fPHYSEQ_phyla_melt$labels_2, "Low abundant phyla", after = 4)
fPHYSEQ_phyla_melt <- fPHYSEQ_phyla_melt[order(fPHYSEQ_phyla_melt$group),]
fPHYSEQ_phyla_melt <- fPHYSEQ_phyla_melt[order(fPHYSEQ_phyla_melt$Time),]
fPHYSEQ_phyla_melt <- fPHYSEQ_phyla_melt[order(fPHYSEQ_phyla_melt$Experiment),]

fPHYSEQ_phyla_melt$SampleID2 <- gsub("_(2|5)W_", "", fPHYSEQ_phyla_melt$SampleID)
levels <- unique(fPHYSEQ_phyla_melt$SampleID2)
fPHYSEQ_phyla_melt$SampleID2 <- factor(fPHYSEQ_phyla_melt$SampleID2, levels=levels)
fPHYSEQ_phyla_melt$Time <- as.factor(fPHYSEQ_phyla_melt$Time)
levels(fPHYSEQ_phyla_melt$Time) <- c("2W", "5W")
levels(fPHYSEQ_phyla_melt$Experiment) <- c("Experiment 1", "Experiment 2")


#plot
fig2.2 <- ggplot(fPHYSEQ_phyla_melt, aes_string(x="SampleID2", y="Abundance", fill="labels_2")) + 
  geom_bar(stat="identity")+
  xlab("") + 
  ylab("Relative abundance [%]") +
  scale_colour_manual(values=fcol_class)+
  scale_fill_manual(values=fcol_class) +
  guides(fill=guide_legend(title="Phylum"))+
  facet_grid(Time ~ Experiment, scale="free_x", space="free_x") +
  geom_text(data=fPHYSEQ_phyla_melt, aes(label=paste0(Herbicide, " ", Application), y=-15), angle=90, check_overlap=T, size=3, hjust = 0.25) + 
  theme_bw()+
  ggtitle("Fungi: Taxonomy") +
  theme(axis.text.x=element_blank(), axis.ticks.x=element_blank())

#png("figures/fig2.2.png",10000,4500, res=600)
print(fig2.2)
#dev.off()

```

\pagebreak

## Effect of all factors on phyla abundances

We test if there is any difference between the phyla abundances between the experiments, the two timepoints, the modes of application, herbicide treatments, any of their interactions or technical lab artefacts (the DNA extraction and PCR amplification was done on two plates) by performing a PERMNOVA (permutations = 999).

### Bacteria

```{r bac PERMANOVA, echo=F, warning=F, message=F}

### BACTERIA

#get phyla abu per sample
bphyla_sum <- aggregate(x=bPHYSEQ_phyla_melt$Abundance, by=list(bPHYSEQ_phyla_melt$labels_2, bPHYSEQ_phyla_melt$Sample), FUN=sum) %>%
  tidyr::pivot_wider(names_from = Group.2, values_from = x, ) %>%
  as.data.frame()
rownames(bphyla_sum) <- bphyla_sum$Group.1
bphyla_sum <- bphyla_sum[,colnames(bphyla_sum) != "Group.1"]

#should be 100% per sample
#colSums(bphyla_sum)

#PERMANOVA

#Experiment 1
bDESIGN_exp1 <- droplevels(bDESIGN[bDESIGN$Experiment==1,])
bphyla_sum_exp1 <- bphyla_sum[,colnames(bphyla_sum) %in% bDESIGN_exp1$Label]
  
bdist_exp1 <- vegdist(t(bphyla_sum_exp1), method="bray")
bdist_paov_exp1 <- adonis2(bdist_exp1 ~ Herbicide * Application * Time + Plate, data=bDESIGN_exp1)
pander(bdist_paov_exp1, caption="Bacteria PERMANOVA: Experiment 1")

#Experiment 2
bDESIGN_exp2 <- droplevels(bDESIGN[bDESIGN$Experiment==2,])
bphyla_sum_exp2 <- bphyla_sum[,colnames(bphyla_sum) %in% bDESIGN_exp2$Label]
  
bdist_exp2 <- vegdist(t(bphyla_sum_exp2), method="bray")
bdist_paov_exp2 <- adonis2(bdist_exp2 ~ Herbicide * Application * Time + Plate, data=bDESIGN_exp2)
pander(bdist_paov_exp2, caption="Bacteria PERMANOVA: Experiment 1")

```

\pagebreak

### Fungi

```{r fun PERMANOVA, echo=F, warning=F, message=F}

### FUNGI

#get phyla abu per sample
fphyla_sum <- aggregate(x=fPHYSEQ_phyla_melt$Abundance, by=list(fPHYSEQ_phyla_melt$labels_2, fPHYSEQ_phyla_melt$Sample), FUN=sum) %>%
  pivot_wider(names_from = Group.2, values_from = x, ) %>%
  as.data.frame()
rownames(fphyla_sum) <- fphyla_sum$Group.1
fphyla_sum <- fphyla_sum[,colnames(fphyla_sum) != "Group.1"]

#should be 100% per sample
#colSums(bphyla_sum)

#Experiment 1
fDESIGN_exp1 <- droplevels(fDESIGN[fDESIGN$Experiment==1,])
fphyla_sum_exp1 <- fphyla_sum[,colnames(fphyla_sum) %in% fDESIGN_exp1$Label]
  
fdist_exp1 <- vegdist(t(fphyla_sum_exp1), method="bray")
fdist_paov_exp1 <- adonis2(fdist_exp1 ~ Herbicide * Application * Time + Plate, data=fDESIGN_exp1)
pander(fdist_paov_exp1, caption="Fungi PERMANOVA: Experiment 1")

#Experiment 2
fDESIGN_exp2 <- droplevels(fDESIGN[fDESIGN$Experiment==2,])
fphyla_sum_exp2 <- fphyla_sum[,colnames(fphyla_sum) %in% fDESIGN_exp2$Label]
  
fdist_exp2 <- vegdist(t(fphyla_sum_exp2), method="bray")
fdist_paov_exp2 <- adonis2(fdist_exp2 ~ Herbicide * Application * Time + Plate, data=fDESIGN_exp2)
pander(fdist_paov_exp2, caption="Fungi PERMANOVA: Experiment 1")

#export for paper

#combine tables
df1 <- cbind(Taxa="Bacteria", Experiment="Experiment1", Factor=rownames(bdist_paov_exp1), bdist_paov_exp1)
df2 <- cbind(Taxa="Bacteria", Experiment="Experiment2", Factor=rownames(bdist_paov_exp2), bdist_paov_exp2)
df3 <- cbind(Taxa="Fungi", Experiment="Experiment1", Factor=rownames(fdist_paov_exp1), fdist_paov_exp1)
df4 <- cbind(Taxa="Fungi", Experiment="Experiment2", Factor=rownames(fdist_paov_exp2), fdist_paov_exp2)
df_all <- rbind(df1, df2, df3, df4)
rownames(df_all) <- NULL
rm(df1, df2, df3, df4)

#save
#write.csv(df_all, "paper_tables/phyla_abundance.csv", row.names = F)

```

**Conclusion:** The relative phyla abundances change over time for bacteria and fungi.

```{r export RDA files, , echo=F, message=F, warning=F}

# create directory
dir.create("interim")
dir.create("interim/02_Taxa_Analysis")

## set output directory
setwd("interim/02_Taxa_Analysis")

#save objects needed in the following scripts as RDA
saveRDS(bPHYSEQ, "bPHYSEQ.RDS")
saveRDS(bPHYSEQ_phyla_melt, "bPHYSEQ_phyla_melt.RDS")

saveRDS(fPHYSEQ, "fPHYSEQ.RDS")
saveRDS(fPHYSEQ_phyla_melt, "fPHYSEQ_phyla_melt.RDS")

```

\pagebreak
