==> Microbiome_Analysis/5_stats/05_Taxa_Response.Rmd <==
---
title: "Herbicide - 05 BX Taxa Response"
author: "Jan Waelchli"
geometry: margin=2cm
output:
  pdf_document:
    toc: yes
    toc_depth: 3
---
   
```{r setup, include=FALSE}

##clear the object from memory
rm(list=ls())

#knitr settings
knitr::opts_chunk$set(echo=TRUE, fig.align="center")
options(tinytex.verbose = TRUE)

#set seed
set.seed(100)

## set source to file location
library(rstudioapi)
setwd(dirname(getActiveDocumentContext()$path))

## load functions
source("functions/functions.R")
functions(path="functions/")

## install (if necessary) and load libraries
libraries()

```

```{r paths, echo=F, message=F, warning=F}

# Import

#set wd to RDS files
setwd("interim/01_Import_Normalization")
#Import RDS files
bDESIGN <- readRDS("bDESIGN.RDS")
bDAT <- readRDS("bDAT.RDS")
bDAT_rare <- readRDS("bDAT_rare.RDS")
bTAX <- readRDS("bTAX.RDS")
b_rare <- readRDS("b_rare.RDS")

fDESIGN <- readRDS("fDESIGN.RDS")
fDAT <- readRDS("fDAT.RDS")
fDAT_rare <- readRDS("fDAT_rare.RDS")
fTAX <- readRDS("fTAX.RDS")
f_rare <- readRDS("f_rare.RDS")

level_cols_Herbicide <- readRDS("level_cols_Herbicide.RDS")
level_cols_Application <- readRDS("level_cols_Application.RDS")

bPHYSEQ <- readRDS("../02_Taxa_Analysis/bPHYSEQ.RDS")
fPHYSEQ <- readRDS("../02_Taxa_Analysis/fPHYSEQ.RDS")

```

# Taxa response

Did a core of microbial taxa respond to herbicides? We searched sensitive ASVs – ASVs being differential abundant between two treatments. We followed the same structure as before and answered the 2 questions:

* **Q1: Are there sensitive ASVs between the herbicide applications?**
* **Q2: Are there sensitive ASVs between the herbicide treatments?**

## Method

To answer the questions we used the package edgeR (Robinson et al., 2010). We fitted a negative binomial generalized log-linear model to the number of reads for each ASV, conducted a likelihood ratio test for the given coefficient contrasts and controlled the family-wise error rate with a Bonferroni-Holms correction.

```{r function herbicide time, warning=F, echo=F}

get_Herbicide_Time_Experiment <- function(DESIGN, DAT, TAX, Herbicide, Time, Experiment=c(1,2)){
  
  #DESIGN
  #subsetting to compartment samples
  DESIGN_Herbicide_Time_Experiment <- DESIGN [DESIGN$Time %in% Time & DESIGN$Herbicide %in% Herbicide & DESIGN$Experiment %in% Experiment,]
  DESIGN_Herbicide_Time_Experiment <- droplevels(DESIGN_Herbicide_Time_Experiment)
  
  #ASV table
  DAT_Herbicide_Time_Experiment <- DAT[,DESIGN_Herbicide_Time_Experiment$Label]
  
  #taxa table
  TAX_Herbicide_Time_Experiment <- TAX[rownames(TAX) %in% rownames(DAT_Herbicide_Time_Experiment),]
  
  return(list(DESIGN_Herbicide_Time_Experiment, DAT_Herbicide_Time_Experiment, TAX_Herbicide_Time_Experiment))
}

```

\pagebreak

## Application effect

**Q1: Are there sensitive ASVs between the herbicide applications?**  
The herbicide was applied directly on soil or on weeds. We checked for different abundant ASVs after application herbicide.

```{r function sens Application, warning=F, echo=F}

sensitive_to_application <- function(DESIGN, DAT, TAX, Herbicide, Time, Experiment, threshold.filter=7, species="bacteria"){
  
  #subset datasets
  temp <-   get_Herbicide_Time_Experiment(DESIGN, DAT, TAX, Herbicide, Time, Experiment)
  DES_sub <- temp[[1]]
  DAT_sub <- temp[[2]]
  TAX_sub <- temp[[3]]
  rm(temp)

  ### edgeR statistics
  #create EDGER object, no TMM done
  EDGER <- edgeR.object.TMM(dat=DAT_sub, threshold.filter=threshold.filter, groups=DES_sub$Application, tax=TAX_sub) #custom function
  ## fit model, estimate dispersion (on edgeR object) using
  # custom function fit.model.edgeR() - source("functions/fun_norm_edgeR_obj.R") 
  EDGER_fit <- fit.model.edgeR(EDGER, DES_sub$Application)

  #glmLRT - weeds vs soil
  contrasts <- makeContrasts(weeds - soil, levels=DES_sub$Application) 
  EDGER_fit_lrt <- glmLRT(EDGER_fit, contrast=contrasts)
  EDGER_fit_lrt_topTags <- topTags(EDGER_fit_lrt, n=Inf, p.value=1)  # all_exp1 stat tests
  EDGER_fit_lrt_table <- topTags(EDGER_fit_lrt, n=nrow(EDGER_fit$counts), adjust.method="BH", sort.by="PValue", p.value=0.05)
  EDGER_fit_lrt_table_ASVs <- rownames(EDGER_fit_lrt_table)
  EDGER_fit_lrt_table_ASVs_percent <- round(100 / nrow(EDGER_fit$counts) * length(EDGER_fit_lrt_table_ASVs), 1)

  # summary of differentiall_exp1y abundant ASVs
  EDGER_summary <- t(summary(decideTestsDGE(EDGER_fit_lrt, p.value=0.05)))
  colnames(EDGER_summary) <- c("higher in soil", "unchanged", "higher in weeds")
  rownames(EDGER_summary) <- NULL

  # # Identification of differentiall_exp1y abundant ASVs
  lrt_ASVs <- rownames(topTags(EDGER_fit_lrt, n=nrow(DAT_sub), adjust.method="BH", sort.by="PValue", p.value=0.05))
  changed <- cbind(TAX[lrt_ASVs, ], EDGER_fit_lrt_table[["table"]][["logFC"]])
  if(species=="bacteria"){try(colnames(changed)[10] <- "logFC", silent = T)}
  if(species=="fungi"){try(colnames(changed)[11] <- "logFC", silent = T)}
  
  try(changed$logFC <- round(changed$logFC, 3))
  colnames(changed)[colnames(changed)=="logFC"] <- "logFC weeds:soil"
  
  return(list(EDGER_summary, changed, EDGER_fit_lrt_table_ASVs, EDGER_fit_lrt_topTags, lrt_ASVs, EDGER_fit_lrt_table))

}

```

```{r function cRA Application, warning=F, echo=F}

cRA_Application <- function(DESIGN, DAT_rare, Herbicide, Time, Experiment, changed){
  
  ##samples
  samples_weeds <- DESIGN$Label[which(DESIGN$Herbicide %in% Herbicide & DESIGN$Time %in% Time & DESIGN$Experiment %in% Experiment) &  DESIGN$Application %in% "weeds"]
  samples_soil <- DESIGN$Label[which(DESIGN$Herbicide %in% Herbicide & DESIGN$Time %in% Time & DESIGN$Experiment %in% Experiment) &  DESIGN$Application %in% "soil"]
  
  #subsetting and normalization
  DAT_rare_weeds <- DAT_rare[, samples_weeds ]
  DAT_rare_weeds <- t(t(DAT_rare_weeds)/colSums(DAT_rare_weeds)) * 100 #normalize (each sample = 100%)
  DAT_rare_weeds <- DAT_rare_weeds[ rowSums(DAT_rare_weeds) > 0,] #remove empty ASV
  
  DAT_rare_soil <- DAT_rare[, samples_soil ]
  DAT_rare_soil <- t(t(DAT_rare_soil)/colSums(DAT_rare_soil)) * 100
  DAT_rare_soil <- DAT_rare_soil[ rowSums(DAT_rare_soil) > 0,]
  
  ## calculation of means and SE
  # weeds and sorting for rank abundance 
  DAT_rare_weeds_MEAN <- apply(DAT_rare_weeds, 1, mean) #mean over all_exp1 samples for each ASVs
  DAT_rare_weeds_MEAN <- sort(DAT_rare_weeds_MEAN, decr=T)
  
  # soil
  DAT_rare_soil_MEAN <- apply(DAT_rare_soil, 1, mean)[names(DAT_rare_weeds_MEAN)]
  
  # summary
  DAT_rare_MEANs <- as.data.frame(cbind(DAT_rare_weeds_MEAN, DAT_rare_soil_MEAN))
  colnames(DAT_rare_MEANs) <- c("weeds","soil")
  
  #changed
  DAT_rare_MEANs$changed <- F
  DAT_rare_MEANs$changed[rownames(DAT_rare_MEANs) %in% rownames(changed)] <- T
  
  #cumulative relative abundance of changed ASVs between weeds and soil
  cRA_weeds <- sum(DAT_rare_MEANs$weeds[DAT_rare_MEANs$changed])
  cRA_soil <- sum(DAT_rare_MEANs$soil[DAT_rare_MEANs$changed])
  
  #sort by ASV rank
  changed$rank <- as.numeric(gsub("ASV", "", changed$ASV_ID))
  changed <- changed[order(changed$rank),]
  
  DAT_rare_MEANs$rank <- as.numeric(gsub("ASV", "", rownames(DAT_rare_MEANs)))
  DAT_rare_MEANs <- DAT_rare_MEANs[order(DAT_rare_MEANs$rank),] #sort by ASV rank

  #add abundance to changed-data
  changed$rAbu_weeds <- DAT_rare_MEANs[rownames(DAT_rare_MEANs) %in% changed$ASV_ID, colnames(DAT_rare_MEANs)=="weeds"]
  changed$rAbu_soil <- DAT_rare_MEANs[rownames(DAT_rare_MEANs) %in% changed$ASV_ID, colnames(DAT_rare_MEANs)=="soil"]
  changed <- changed[order(changed$rAbu_weeds, decreasing = T),] #sort by abu
  try(changed$rAbu_weeds <- paste0(round(changed$rAbu_weeds,3), "%"), silent = T)
  try(changed$rAbu_soil <- paste0(round(changed$rAbu_soil,3), "%"), silent = T)

  return(list(cRA_weeds, cRA_soil, changed))

}

```

### Bacteria

```{r bac application, warning=F, echo=F}

#exp1

temp <-   sensitive_to_application(bDESIGN, bDAT, bTAX, c("ctr","gly","tb"), c(2,5), 1, species = "bacteria")
bEDGER_summary_exp1 <- temp[[1]]
bchanged_exp1 <- temp[[2]]
bEDGER_fit_lrt_table_ASVs_exp1 <- temp[[3]]
bEDGER_fit_lrt_topTags_exp1 <- temp[[4]]
lrt_bASVs_exp1 <- temp[[5]]
bEDGER_fit_lrt_table_exp1 <- temp[[6]]

temp <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","gly","tb"), c(2,5), 1, bchanged_exp1)
bCRA_exp1 <- temp[[1]]
bchanged_exp1 <- temp[[3]]

pander(bEDGER_summary_exp1[,c(3,2,1)], justify='left', caption = "Bacteria: Application effect in experiment 1") 
print(paste0(round(bCRA_exp1, 2), "%", " of the bacterial abundance changed between the different application targets."))
bchanged_exp1_out <- bchanged_exp1[, c(2,3,10,12, 13)]
kbl(bchanged_exp1_out[1:min(10, nrow(bchanged_exp1)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")

#write.csv(bchanged_exp1_out, "paper_tables/bchanged_exp1_application.csv")

#exp2

temp <-   sensitive_to_application(bDESIGN, bDAT, bTAX, c("ctr","gly","tb"), c(2,5), 2, species = "bacteria")
bEDGER_summary_exp2 <- temp[[1]]
bchanged_exp2 <- temp[[2]]
bEDGER_fit_lrt_table_ASVs_exp2 <- temp[[3]]
bEDGER_fit_lrt_topTags_exp2 <- temp[[4]]
lrt_bASVs_exp2 <- temp[[5]]
bEDGER_fit_lrt_table_exp2 <- temp[[6]]

temp <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","gly","tb"), c(2,5), 2, bchanged_exp2)
bCRA_exp2 <- temp[[1]]
bchanged_exp2 <- temp[[3]]

pander(bEDGER_summary_exp2[,c(3,2,1)], justify='left', caption = "Bacteria: Application effect in experiment 2") 
print(paste0(round(bCRA_exp2, 2), "%", " of the bacterial abundance changed between the different application targets."))
bchanged_exp2_out <- bchanged_exp2[, c(2,3,10,12, 13)]
kbl(bchanged_exp2_out[1:min(10, nrow(bchanged_exp2)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")
#write.csv(bchanged_exp2_out, "paper_tables/bchanged_exp2_application.csv")

```

In experiment 1 were `r paste(length(bEDGER_fit_lrt_table_ASVs_exp1))` bacterial ASVs sensitive to different application targets while we found in experiment 2 `r paste(length(bEDGER_fit_lrt_table_ASVs_exp2))` sensitive bacterial ASVs.

\vspace{5mm}

### Figure 6.1 | Bacteria: MA & venn

\vspace{5mm}

```{r MA and venn bacteria application, echo=F, warning=F, message=F, eval=T, fig.height=8, fig.width=9}

#colors
col_weeds_up <- level_cols_Application[names(level_cols_Application) == "weeds"] %>% as.character()
col_soil_up <- level_cols_Application[names(level_cols_Application) == "soil"] %>% as.character()

#create bacteria df
df1 <- bEDGER_fit_lrt_topTags_exp1[["table"]]
df1$experiment <- "Experiment 1"
df2 <- bEDGER_fit_lrt_topTags_exp2[["table"]]
df2$experiment <- "Experiment 2"
df <- rbind(df1, df2)
rm(df1, df2)

#get up and down-regulated ASVs
df$cols <- "NS"
df$logFC <- df$logFC * (-1) #change from hc:ctr to ctr:hc #change from soil:weeds to weeds:soil
df$cols[df$FDR < 0.05 & df$logFC > 0] <- "higher in weeds"
df$cols[df$FDR < 0.05 & df$logFC < 0] <- "higher in soil"

#sort to plot sign at the end
df <- df[order(df$FDR, decreasing = T),]

#factors of col plot
df$cols <- factor(df$cols, levels=c("higher in weeds", "higher in soil", "NS"))

#plot
fig6.1.1 <- ggplot(df, aes(x=logCPM, y=logFC))+
              geom_point(aes(col=cols), alpha=0.8, size=2)+
              scale_color_manual(values=c(col_weeds_up,col_soil_up, "grey"))+
              facet_wrap(~ experiment)+
              theme_bw()+
              ylab("log2 fold change")+
              xlab("log2 mean abundance")


#VENN

#higher in ctr
#sensitive bacteria
exp1_ASV_higher <- bchanged_exp1$ASV_ID[bchanged_exp1$`logFC weeds:soil`>0]
exp2_ASV_higher <- bchanged_exp2$ASV_ID[bchanged_exp2$`logFC weeds:soil`>0]
list_venn <- list(exp1_higher = exp1_ASV_higher, exp2_higher = exp2_ASV_higher)
#relative abundance
bCRA_higher_exp1 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","gly","tb"), c(2,5), 2, bchanged_exp1[bchanged_exp1$`logFC weeds:soil` > 0,])[[1]]
bCRA_higher_exp2 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","gly","tb"), c(2,5), 2, bchanged_exp2[bchanged_exp2$`logFC weeds:soil` > 0,])[[1]]
#labels
labels <- c(paste0("exp 1\ncRA: ", round(bCRA_higher_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(bCRA_higher_exp2, 2), "%"))
names(list_venn) <- labels
#plot
bvenn_h <- ggvenn(list_venn, fill_color = c(col_weeds_up, col_weeds_up), stroke_size = 0.5, set_name_size = 3, show_percentage = F) +
  facet_wrap(~"higher in weeds")+ #to set title
  theme_bw()+ #to get grey title box
  theme(axis.title.x=element_blank(),axis.text.x=element_blank(),axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),axis.text.y=element_blank(),axis.ticks.y=element_blank(),
        line = element_blank()) #to remove axes and axes lines
  

list_venn1 <- list_venn

#lower in ctr
#sensitive bacteria
exp1_ASV_lower <- bchanged_exp1$ASV_ID[bchanged_exp1$`logFC weeds:soil`<0]
exp2_ASV_lower <- bchanged_exp2$ASV_ID[bchanged_exp2$`logFC weeds:soil`<0]
list_venn <- list(exp1_lower = exp1_ASV_lower, exp2_lower = exp2_ASV_lower)
#relative abundance
bCRA_lower_exp1 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","gly","tb"), c(2,5), 2, bchanged_exp1[bchanged_exp1$`logFC weeds:soil` < 0,])[[1]]
bCRA_lower_exp2 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","gly","tb"), c(2,5), 2, bchanged_exp2[bchanged_exp2$`logFC weeds:soil` < 0,])[[1]]
#labels
labels <- c(paste0("exp 1\ncRA: ", round(bCRA_lower_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(bCRA_lower_exp2, 2), "%"))
names(list_venn) <- labels
#plot
bvenn_l <- ggvenn(list_venn, fill_color = c(col_soil_up, col_soil_up), stroke_size = 0.5, set_name_size = 3, show_percentage = F) +
  facet_wrap(~"higher in soil")+ #to set title
  theme_bw()+ #to get grey title box
  theme(axis.title.x=element_blank(),axis.text.x=element_blank(),axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),axis.text.y=element_blank(),axis.ticks.y=element_blank(),
        line = element_blank()) #to remove axes and axes lines

list_venn2 <- list_venn

#combine plots
fig6.1.2 <- plot_grid(
                 #NULL,
                 bvenn_h + theme(legend.position="none", plot.margin = unit(c(0, 0.1, 0, 0.1), "cm")),
                 bvenn_l+ theme(legend.position="none", plot.margin = unit(c(0, 0.2, 0, 0.1), "cm")) + labs(colour=""),
                 align = "hv",
                 #rel_widths = c(0.2,1,1),
                 nrow=1)

fig6.1 <- plot_grid(fig6.1.1 + theme(legend.position="none"),
                 fig6.1.2+ theme(legend.position="none") + labs(colour=""),
                 #labels=c("A","B"),
                 align = "hv",
                 nrow=2)

#saveRDS(fig6.1, "paper_figs/RDS/fig6.1.RDS")

#add legend
legend <- get_legend(fig6.1.1 + theme(legend.position="bottom") + labs(col="Application"))
fig6.1 <- plot_grid(fig6.1, legend, nrow=2, rel_heights=c(1, .1))

#print
fig6.1

```

\pagebreak

### Fungi

```{r fun application, warning=F, echo=F}

#exp1
temp <-   sensitive_to_application(fDESIGN, fDAT, fTAX, c("ctr","gly","tb"), c(2,5), 1, species = "fungi")
fEDGER_summary_exp1 <- temp[[1]]
fchanged_exp1 <- temp[[2]]
fEDGER_fit_lrt_table_ASVs_exp1 <- temp[[3]]
fEDGER_fit_lrt_topTags_exp1 <- temp[[4]]
lrt_fASVs_exp1 <- temp[[5]]
fEDGER_fit_lrt_table_exp1 <- temp[[6]]

temp <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","gly","tb"), c(2,5), 1, fchanged_exp1)
fCRA_exp1 <- temp[[1]]
fchanged_exp1 <- temp[[3]]

pander(fEDGER_summary_exp1[,c(3,2,1)], justify='left', caption = "Fungi: Application effect in experiment 1") 
print(paste0(round(fCRA_exp1, 2), "%", " of the fungal abundance changed between the different application targets."))
fchanged_exp1_out <- fchanged_exp1[, c(2,3,11,13,14)]
kbl(fchanged_exp1_out[1:min(10, nrow(fchanged_exp1)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")
#write.csv(fchanged_exp1_out, "paper_tables/fchanged_exp1_application.csv")


#exp2
temp <-   sensitive_to_application(fDESIGN, fDAT, fTAX, c("ctr","gly","tb"), c(2,5), 2, species = "fungi")
fEDGER_summary_exp2 <- temp[[1]]
fchanged_exp2 <- temp[[2]]
fEDGER_fit_lrt_table_ASVs_exp2 <- temp[[3]]
fEDGER_fit_lrt_topTags_exp2 <- temp[[4]]
lrt_fASVs_exp2 <- temp[[5]]
fEDGER_fit_lrt_table_exp2 <- temp[[6]]

temp <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","gly","tb"), c(2,5), 2, fchanged_exp2)
fCRA_exp2 <- temp[[1]]
fchanged_exp2 <- temp[[3]]

pander(fEDGER_summary_exp2[,c(3,2,1)], justify='left', caption = "Fungi: Application effect in experiment 2") 
print(paste0(round(fCRA_exp2, 2), "%", " of the fungal abundance changed between the different application targets."))
fchanged_exp2_out <- fchanged_exp2[, c(2,3,11,13,14)]
kbl(fchanged_exp2_out[1:min(10, nrow(fchanged_exp2)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")
#write.csv(fchanged_exp2_out, "paper_tables/fchanged_exp2_application.csv")

```

In experiment 1 were `r paste(length(fEDGER_fit_lrt_table_ASVs_exp1))` fungal ASVs sensitive to different application targets while we found in experiment 2 `r paste(length(fEDGER_fit_lrt_table_ASVs_exp2))` sensitive fungal ASVs.

\vspace{5mm}

### Figure 6.2 | Fungi: MA & venn

\vspace{5mm}

```{r MA and venn fungi application, echo=F, warning=F, message=F, eval=T, fig.height=8, fig.width=9}

#create fungi df
df1 <- fEDGER_fit_lrt_topTags_exp1[["table"]]
df1$experiment <- "Experiment 1"
df2 <- fEDGER_fit_lrt_topTags_exp2[["table"]]
df2$experiment <- "Experiment 2"
df <- rbind(df1, df2)
rm(df1, df2)

#get up and down-regulated ASVs
df$cols <- "NS"
#df$logFC <- df$logFC * (-1) #change from hc:ctr to ctr:hc #change from soil:weeds to weeds:soil
df$cols[df$FDR < 0.05 & df$logFC > 0] <- "higher in weeds"
df$cols[df$FDR < 0.05 & df$logFC < 0] <- "higher in soil"

#sort to plot sign at the end
df <- df[order(df$FDR, decreasing = T),]

#factors of col plot
df$cols <- factor(df$cols, levels=c("higher in weeds", "higher in soil", "NS"))

fig6.2.1 <- ggplot(df, aes(x=logCPM, y=logFC))+
              geom_point(aes(col=cols), alpha=0.8, size=2)+
              scale_color_manual(values=c(col_weeds_up,col_soil_up, "grey"))+
              facet_wrap(~ experiment)+
              theme_bw()+
              ylab("log2 fold change")+
              xlab("log2 mean abundance")


#VENN

#higher in ctr
#sensitive fungi
exp1_ASV_higher <- fchanged_exp1$ASV_ID[fchanged_exp1$`logFC weeds:soil`>0]
exp2_ASV_higher <- fchanged_exp2$ASV_ID[fchanged_exp2$`logFC weeds:soil`>0]
list_venn <- list(exp1_higher = exp1_ASV_higher, exp2_higher = exp2_ASV_higher)
#relative abundance
fCRA_higher_exp1 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","gly","tb"), c(2,5), 2, fchanged_exp1[fchanged_exp1$`logFC weeds:soil` > 0,])[[1]]
fCRA_higher_exp2 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","gly","tb"), c(2,5), 2, fchanged_exp2[fchanged_exp2$`logFC weeds:soil` > 0,])[[1]]
#labels
labels <- c(paste0("exp 1\ncRA: ", round(fCRA_higher_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(fCRA_higher_exp2, 2), "%"))
names(list_venn) <- labels
#plot
fvenn_h <- ggvenn(list_venn, fill_color = c(col_weeds_up, col_weeds_up), stroke_size = 0.5, set_name_size = 3, show_percentage = F) +
  facet_wrap(~"higher in weeds")+ #to set title
  theme_bw()+ #to get grey title box
  theme(axis.title.x=element_blank(),axis.text.x=element_blank(),axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),axis.text.y=element_blank(),axis.ticks.y=element_blank(),
        line = element_blank()) #to remove axes and axes lines
  

#lower in ctr
#sensitive fungi
exp1_ASV_lower <- fchanged_exp1$ASV_ID[fchanged_exp1$`logFC weeds:soil`<0]
exp2_ASV_lower <- fchanged_exp2$ASV_ID[fchanged_exp2$`logFC weeds:soil`<0]
list_venn <- list(exp1_lower = exp1_ASV_lower, exp2_lower = exp2_ASV_lower)
#relative abundance
fCRA_lower_exp1 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","gly","tb"), c(2,5), 2, fchanged_exp1[fchanged_exp1$`logFC weeds:soil` < 0,])[[1]]
fCRA_lower_exp2 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","gly","tb"), c(2,5), 2, fchanged_exp2[fchanged_exp2$`logFC weeds:soil` < 0,])[[1]]
#labels
labels <- c(paste0("exp 1\ncRA: ", round(fCRA_lower_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(fCRA_lower_exp2, 2), "%"))
names(list_venn) <- labels
#plot
fvenn_l <- ggvenn(list_venn, fill_color = c(col_soil_up, col_soil_up), stroke_size = 0.5, set_name_size = 3, show_percentage = F) +
  facet_wrap(~"higher in soil")+ #to set title
  theme_bw()+ #to get grey title box
  theme(axis.title.x=element_blank(),axis.text.x=element_blank(),axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),axis.text.y=element_blank(),axis.ticks.y=element_blank(),
        line = element_blank()) #to remove axes and axes lines

#combine plots
fig6.2.2 <- plot_grid(
                 #NULL,
                 fvenn_h + theme(legend.position="none", plot.margin = unit(c(0, 0.1, 0, 0.1), "cm")),
                 fvenn_l+ theme(legend.position="none", plot.margin = unit(c(0, 0.2, 0, 0.1), "cm")) + labs(colour=""),
                 align = "hv",
                 #rel_widths = c(0.2,1,1),
                 nrow=1)

fig6.2 <- plot_grid(fig6.2.1 + theme(legend.position="none"),
                 fig6.2.2+ theme(legend.position="none") + labs(colour=""),
                 #labels=c("A","B"),
                 align = "hv",
                 nrow=2)

#saveRDS(fig6.2, "paper_figs/RDS/fig6.2.RDS")

#add legend
legend <- get_legend(fig6.2.1 + theme(legend.position="bottom") + labs(col="Application"))
fig6.2 <- plot_grid(fig6.2, legend, nrow=2, rel_heights=c(1, .1))

#print
fig6.2


```

\vspace{5mm}

**Conclusion:** Different application targets altered bacterial and fungal community members in their relative abundance. The majority of sensitive ASVs are not abundant in the rhizosphere, that's why only small_exp1 percentages were changed due to application herbicides. Most of the changed ASVs were only sensitive either in experiment 1 or experiment 2.

\pagebreak

## Herbicide effect

**Q2: Are there sensitive ASVs between the herbicide treatments?**  
We treated samples either with water (control), glyphosate or terbuthylazine. We checked for both experiments for differences in the ASV abundance by comparing the herbicides against the water control.

```{r function Herbicide, warning=F, echo=F}

sensitive_to_herbicide <- function(DESIGN, DAT, TAX, Herbicide, Time, Experiment, threshold.filter=7, species="bacteria"){
  
  #subset datasets
  temp <-   get_Herbicide_Time_Experiment(DESIGN, DAT, TAX, Herbicide, Time, Experiment)
  DES_sub <- temp[[1]]
  DAT_sub <- temp[[2]]
  TAX_sub <- temp[[3]]
  rm(temp)
  
  ### edgeR statistics
  ## Preparing EdgeR object, filtering and normalization (TMM) using
  # custom function edgeR.object.TMM() - source("functions/fun_norm_edgeR_obj.R") 
  EDGER <- edgeR.object.TMM(dat=DAT_sub, threshold.filter=threshold.filter, groups=DES_sub$Herbicide, tax=TAX_sub)
  ## fit model, estimate dispersion (on edgeR object) using
  # custom function fit.model.edgeR() - source("functions/fun_norm_edgeR_obj.R") 
  EDGER_fit <- fit.model.edgeR(EDGER, DES_sub$Herbicide)

  #glmLRT - weeds vs soil
  if("gly" %in% Herbicide){contrasts <- makeContrasts(ctr - gly, levels=DES_sub$Herbicide)}
  if("tb" %in% Herbicide){contrasts <- makeContrasts(ctr - tb, levels=DES_sub$Herbicide)}
  EDGER_fit_lrt <- glmLRT(EDGER_fit, contrast=contrasts)
  EDGER_fit_lrt_topTags <- topTags(EDGER_fit_lrt, n=Inf, p.value=1)  # all_exp1 stat tests
  EDGER_fit_lrt_table <- topTags(EDGER_fit_lrt, n=nrow(EDGER_fit$counts), adjust.method="BH", sort.by="PValue", p.value=0.05)
  EDGER_fit_lrt_table_ASVs <- rownames(EDGER_fit_lrt_table)
  EDGER_fit_lrt_table_ASVs_percent <- round(100 / nrow(EDGER_fit$counts) * length(EDGER_fit_lrt_table_ASVs), 1)

  # summary of differentiall_exp1y abundant ASVs
  EDGER_summary <- t(summary(decideTestsDGE(EDGER_fit_lrt, p.value=0.05)))
  if("gly" %in% Herbicide){colnames(EDGER_summary) <- c("higher in gly", "unchanged", "lower in gly")}
  if("tb" %in% Herbicide){colnames(EDGER_summary) <- c("higher in tb", "unchanged", "lower in tb")}
  rownames(EDGER_summary) <- NULL

  # # Identification of differentiall_exp1y abundant ASVs
  lrt_ASVs <- rownames(topTags(EDGER_fit_lrt, n=nrow(DAT_sub), adjust.method="BH", sort.by="PValue", p.value=0.05))
  changed <- cbind(TAX[lrt_ASVs, ], EDGER_fit_lrt_table[["table"]][["logFC"]])
  if(species=="bacteria"){try(colnames(changed)[10] <- "logFC", silent = T)}
  if(species=="fungi"){try(colnames(changed)[11] <- "logFC", silent = T)}
  
  try(changed$logFC <- round(changed$logFC, 3))
  colnames(changed)[colnames(changed)=="logFC"] <- "logFC ctr:hc"

  return(list(EDGER_summary, changed, EDGER_fit_lrt_table_ASVs, EDGER_fit_lrt_topTags, lrt_ASVs, EDGER_fit_lrt_table))

  
}

```

```{r function cRA Herbicide, warning=F, echo=F}

cRA_Herbicide <- function(DESIGN, DAT_rare, Herbicide, Time, Experiment, changed){
  
  ##samples
  samples_hc <- DESIGN$Label[which(DESIGN$Herbicide %in% Herbicide & DESIGN$Time %in% Time & DESIGN$Experiment %in% Experiment)]
  samples_ctr <- DESIGN$Label[which(DESIGN$Herbicide %in% "ctr" & DESIGN$Time %in% Time & DESIGN$Experiment %in% Experiment)]
  
  #subsetting and normalization
  DAT_rare_hc <- DAT_rare[, samples_hc ]
  DAT_rare_hc <- t(t(DAT_rare_hc)/colSums(DAT_rare_hc)) * 100 #normalize (each sample = 100%)
  DAT_rare_hc <- DAT_rare_hc[ rowSums(DAT_rare_hc) > 0,] #remove empty ASV
  
  DAT_rare_ctr <- DAT_rare[, samples_ctr ]
  DAT_rare_ctr <- t(t(DAT_rare_ctr)/colSums(DAT_rare_ctr)) * 100
  DAT_rare_ctr <- DAT_rare_ctr[ rowSums(DAT_rare_ctr) > 0,]
  
  ## calculation of means and SE
  # hc and sorting for rank abundance 
  DAT_rare_hc_MEAN <- apply(DAT_rare_hc, 1, mean) #mean over all_exp1 samples for each ASVs
  DAT_rare_hc_MEAN <- sort(DAT_rare_hc_MEAN, decr=T)
  
  # ctr
  DAT_rare_ctr_MEAN <- apply(DAT_rare_ctr, 1, mean)[names(DAT_rare_hc_MEAN)]
  
  # summary
  DAT_rare_MEANs <- as.data.frame(cbind(DAT_rare_hc_MEAN, DAT_rare_ctr_MEAN))
  colnames(DAT_rare_MEANs) <- c("hc","ctr")
  
  #changed
  DAT_rare_MEANs$changed <- F
  DAT_rare_MEANs$changed[rownames(DAT_rare_MEANs) %in% rownames(changed)] <- T
  
  #cumulative relative abundance of changed ASVs between hc and ctr
  cRA_hc <- sum(DAT_rare_MEANs$hc[DAT_rare_MEANs$changed])
  cRA_ctr <- sum(DAT_rare_MEANs$ctr[DAT_rare_MEANs$changed])
  
  #sort by ASV rank
  changed$rank <- as.numeric(gsub("ASV", "", changed$ASV_ID))
  changed <- changed[order(changed$rank),]

  DAT_rare_MEANs$rank <- as.numeric(gsub("ASV", "", rownames(DAT_rare_MEANs)))
  DAT_rare_MEANs <- DAT_rare_MEANs[order(DAT_rare_MEANs$rank),] #sort by ASV rank
  
  #add abundance to changed-data
  changed <- changed[changed$ASV_ID %in% rownames(DAT_rare_MEANs),]#ommit errors if not all_exp1 ASVs in changed are in DAT_rare
  changed$rAbu_ctr <- DAT_rare_MEANs[rownames(DAT_rare_MEANs) %in% changed$ASV_ID,  colnames(DAT_rare_MEANs)=="ctr"]#add abu
  changed$rAbu_hc <- DAT_rare_MEANs[rownames(DAT_rare_MEANs) %in% changed$ASV_ID,  colnames(DAT_rare_MEANs)=="hc"]#add abu
  changed <- changed[order(changed$rAbu_ctr, decreasing = T),] #sort by abu
  try(changed$rAbu_ctr <- paste0(round(changed$rAbu_ctr,3), "%"), silent = T)
  try(changed$rAbu_hc <- paste0(round(changed$rAbu_hc,3), "%"), silent = T)

  return(list(cRA_hc, cRA_ctr, changed))

}

```

### Bacteria

```{r bacteria exp1 herbicide, warning=F, echo=F}

#exp1

#gly
temp <-   sensitive_to_herbicide(bDESIGN, bDAT, bTAX, c("ctr","gly"), c(2,5), 1, species = "bacteria")
bEDGER_summary_gly_exp1 <- temp[[1]]
bchanged_gly_exp1 <- temp[[2]]
bEDGER_fit_lrt_table_ASVs_gly_exp1 <- temp[[3]]
bEDGER_fit_lrt_topTags_gly_exp1 <- temp[[4]]
lrt_bASVs_gly_exp1 <- temp[[5]]
bEDGER_fit_lrt_table_gly_exp1 <- temp[[6]]

temp <- cRA_Herbicide(bDESIGN, bDAT_rare, c("ctr","gly"), c(2,5), 1, bchanged_gly_exp1)
bCRA_gly_exp1 <- temp[[1]]
bchanged_gly_exp1 <- temp[[3]]

pander(bEDGER_summary_gly_exp1[,c(3,2,1)], justify='left', caption = "Bacteria: Glyphosate effect in experiment 1") 
print(paste0(round(bCRA_gly_exp1, 2), "%", " of the bacterial abundance changed after applying glyphosate"))
bchanged_gly_exp1_out <- bchanged_gly_exp1[, c(2,3,10,12,13)]
kbl(bchanged_gly_exp1_out[1:min(10, nrow(bchanged_gly_exp1)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")

#write.csv(bchanged_gly_exp1_out, "paper_tables/bchanged_gly_exp1.csv")

#tb
temp <-   sensitive_to_herbicide(bDESIGN, bDAT, bTAX, c("ctr","tb"), c(2,5), 1, species = "bacteria")
bEDGER_summary_tb_exp1 <- temp[[1]]
bchanged_tb_exp1 <- temp[[2]]
bEDGER_fit_lrt_table_ASVs_tb_exp1 <- temp[[3]]
bEDGER_fit_lrt_topTags_tb_exp1 <- temp[[4]]
lrt_bASVs_tb_exp1 <- temp[[5]]
bEDGER_fit_lrt_table_tb_exp1 <- temp[[6]]

temp <- cRA_Herbicide(bDESIGN, bDAT_rare, c("ctr","tb"), c(2,5), 1, bchanged_tb_exp1)
bCRA_tb_exp1 <- temp[[1]]
bchanged_tb_exp1 <- temp[[3]]

pander(bEDGER_summary_tb_exp1[,c(3,2,1)], justify='left', caption = "Bacteria: Terbuthylazine effect in experiment 1") 
print(paste0(round(bCRA_tb_exp1, 2), "%", " of the bacterial abundance changed after applying terbuthylazine"))
bchanged_tb_exp1_out <- bchanged_tb_exp1[, c(2,3,10,12,13)]
kbl(bchanged_tb_exp1_out[1:min(10, nrow(bchanged_tb_exp1_out)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")

#write.csv(bchanged_tb_exp1_out, "paper_tables/bchanged_tb_exp1.csv")

```

\pagebreak

```{r bacteria exp2 herbicide, warning=F, echo=F}

#exp2

#gly
temp <-   sensitive_to_herbicide(bDESIGN, bDAT, bTAX, c("ctr","gly"), c(2,5), 2, species = "bacteria")
bEDGER_summary_gly_exp2 <- temp[[1]]
bchanged_gly_exp2 <- temp[[2]]
bEDGER_fit_lrt_table_ASVs_gly_exp2 <- temp[[3]]
bEDGER_fit_lrt_topTags_gly_exp2 <- temp[[4]]
lrt_bASVs_gly_exp2 <- temp[[5]]
bEDGER_fit_lrt_table_gly_exp2 <- temp[[6]]

temp <- cRA_Herbicide(bDESIGN, bDAT_rare, c("ctr","gly"), c(2,5), 2, bchanged_gly_exp2)
bCRA_gly_exp2 <- temp[[1]]
bchanged_gly_exp2 <- temp[[3]]

pander(bEDGER_summary_gly_exp2[,c(3,2,1)], justify='left', caption = "Bacteria: Glyphosate effect in experiment 2") 
print(paste0(round(bCRA_gly_exp2, 2), "%", " of the bacterial abundance changed after applying glyphosate"))
bchanged_gly_exp2_out <- bchanged_gly_exp2[, c(2,3,10,12,13)]
kbl(bchanged_gly_exp2_out[1:min(10, nrow(bchanged_gly_exp2_out)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")

#write.csv(bchanged_gly_exp2_out, "paper_tables/bchanged_gly_exp2.csv")

#tb
temp <-   sensitive_to_herbicide(bDESIGN, bDAT, bTAX, c("ctr","tb"), c(2,5), 2, species = "bacteria")
bEDGER_summary_tb_exp2 <- temp[[1]]
bchanged_tb_exp2 <- temp[[2]]
bEDGER_fit_lrt_table_ASVs_tb_exp2 <- temp[[3]]
bEDGER_fit_lrt_topTags_tb_exp2 <- temp[[4]]
lrt_bASVs_tb_exp2 <- temp[[5]]
bEDGER_fit_lrt_table_tb_exp2 <- temp[[6]]

temp <- cRA_Herbicide(bDESIGN, bDAT_rare, c("ctr","tb"), c(2,5), 2, bchanged_tb_exp2)
bCRA_tb_exp2 <- temp[[1]]
bchanged_tb_exp2 <- temp[[3]]

pander(bEDGER_summary_tb_exp2[,c(3,2,1)], justify='left', caption = "Bacteria: Terbuthylazine effect in experiment 2") 
print(paste0(round(bCRA_tb_exp2, 2), "%", " of the bacterial abundance changed after applying terbuthylazine"))
bchanged_tb_exp2_out <- bchanged_tb_exp2[, c(2,3,10,12,13)]
kbl(bchanged_tb_exp2_out[1:min(10, nrow(bchanged_tb_exp2_out)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")

#write.csv(bchanged_tb_exp2_out, "paper_tables/bchanged_tb_exp2.csv")

```

Glyphosate changed the abundance of `r paste(length(bEDGER_fit_lrt_table_ASVs_gly_exp1))` bacterial ASVs in experiment 1 and `r paste(length(bEDGER_fit_lrt_table_ASVs_gly_exp2))` bacterial ASVs in experiment 2. Application terbuthylazine led to changes in abundance for `r paste(length(bEDGER_fit_lrt_table_ASVs_tb_exp1))` bacterial ASVs in experiment 1 and `r paste(length(bEDGER_fit_lrt_table_ASVs_tb_exp2))` bacterial ASVs in experiment 2.

### Figure 7.1 | Bacteria: MA & venn

\vspace{5mm}

```{r MA and venn bacteria herbicide, echo=F, warning=F, message=F, eval=T, fig.height=10, fig.width=6}

#colors
col_ctr_up <- level_cols_Herbicide[names(level_cols_Herbicide) == "ctr"] %>% as.character()
col_gly_up <- level_cols_Herbicide[names(level_cols_Herbicide) == "gly"] %>% as.character()
col_tb_up <- level_cols_Herbicide[names(level_cols_Herbicide) == "tb"] %>% as.character()

#gly
#create bacteria df
df1 <- bEDGER_fit_lrt_topTags_gly_exp1[["table"]]
df1$herbicide <- "gly"
df1$experiment <- "Experiment 1"
df2 <- bEDGER_fit_lrt_topTags_gly_exp2[["table"]]
df2$herbicide <- "gly"
df2$experiment <- "Experiment 2"

df <- rbind(df1, df2)
rm(df1, df2)

#get up and down-regulated ASVs
df$cols <- "NS"
#df$logFC <- df$logFC * (-1) #change from hc:ctr to ctr:hc #change from hc:ctr to ctr:hc
df$cols[df$FDR < 0.05 & df$logFC > 0] <- "higher in ctr"
df$cols[df$FDR < 0.05 & df$logFC < 0] <- "higher in gly"

#sort to plot sign at the end
df <- df[order(df$FDR, decreasing = T),]

#plot
fig7.1.1_gly <- ggplot(df, aes(x=logCPM, y=logFC))+
                  geom_point(aes(col=cols), alpha=0.8, size=2)+
                  scale_color_manual(values=c(col_ctr_up,col_gly_up, "grey"))+
                  facet_wrap(~ experiment)+
                  theme_bw()+
                  ylab("log2 fold change")+
                  xlab("log2 mean abundance")

#tb
#create bacteria df
df1 <- bEDGER_fit_lrt_topTags_tb_exp1[["table"]]
df1$herbicide <- "tb"
df1$experiment <- "Experiment 1"
df2 <- bEDGER_fit_lrt_topTags_tb_exp2[["table"]]
df2$herbicide <- "tb"
df2$experiment <- "Experiment 2"

df <- rbind(df1, df2)
rm(df1, df2)

#get up and down-regulated ASVs
df$cols <- "NS"
#df$logFC <- df$logFC * (-1) #change from hc:ctr to ctr:hc #change from hc:ctr to ctr:hc
df$cols[df$FDR < 0.05 & df$logFC > 0] <- "higher in ctr"
df$cols[df$FDR < 0.05 & df$logFC < 0] <- "higher in tb"

#sort to plot sign at the end
df <- df[order(df$FDR, decreasing = T),]

#plot
fig7.1.1_tb <- ggplot(df, aes(x=logCPM, y=logFC))+
                  geom_point(aes(col=cols), alpha=0.8, size=2)+
                  scale_color_manual(values=c(col_ctr_up,col_tb_up, "grey"))+
                  facet_wrap(~ experiment)+
                  theme_bw()+
                  ylab("log2 fold change")+
                  xlab("log2 mean abundance")

#combine MA plots
fig7.1.1 <- plot_grid(fig7.1.1_gly + theme(legend.position="none", plot.margin = unit(c(0.2, 0.2, 0.2, 0.2), "cm")),
                 fig7.1.1_tb+ theme(legend.position="none", plot.margin = unit(c(0.2, 0.2, 0.2, 0.2), "cm")) + labs(colour=""),
                 #labels=c("A","B"),
                 align = "hv",
                 nrow=2)

#VENN

#higher in ctr
#sensitive bacteria gly
exp1_ASV_higher <- bchanged_gly_exp1$ASV_ID[bchanged_gly_exp1$`logFC ctr:hc`>0]
exp2_ASV_higher <- bchanged_gly_exp2$ASV_ID[bchanged_gly_exp2$`logFC ctr:hc`>0]
list_venn_gly <- list(exp1_higher = exp1_ASV_higher, exp2_higher = exp2_ASV_higher)
#relative abundance
bCRA_higher_gly_exp1 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","gly"), c(2,5), 1, bchanged_gly_exp1[bchanged_gly_exp1$`logFC ctr:hc` > 0,])[[1]]
bCRA_higher_gly_exp2 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","gly"), c(2,5), 2, bchanged_gly_exp2[bchanged_gly_exp2$`logFC ctr:hc` > 0,])[[1]]
#sensitive bacteria tb
exp1_ASV_higher <- bchanged_tb_exp1$ASV_ID[bchanged_tb_exp1$`logFC ctr:hc`>0]
exp2_ASV_higher <- bchanged_tb_exp2$ASV_ID[bchanged_tb_exp2$`logFC ctr:hc`>0]
list_venn_tb <- list(exp1_higher = exp1_ASV_higher, exp2_higher = exp2_ASV_higher)
#relative abundance
bCRA_higher_tb_exp1 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","tb"), c(2,5), 1, bchanged_tb_exp1[bchanged_tb_exp1$`logFC ctr:hc` > 0,])[[1]]
bCRA_higher_tb_exp2 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","tb"), c(2,5), 2, bchanged_tb_exp2[bchanged_tb_exp2$`logFC ctr:hc` > 0,])[[1]]

#labels
list_venn <- c(list_venn_gly, list_venn_tb)
labels <- c(paste0("exp 1\ncRA: ", round(bCRA_higher_gly_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(bCRA_higher_gly_exp2, 2), "%"),
            paste0("exp 1\ncRA: ", round(bCRA_higher_tb_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(bCRA_higher_tb_exp2, 2), "%"))
names(list_venn) <- labels

#plot
bvenn_h <- ggvenn(list_venn, fill_color = c(col_gly_up, col_gly_up, col_tb_up, col_tb_up), stroke_size = 0.5, set_name_size = 2, show_percentage = F) +
  facet_wrap(~"higher in ctr")+ #to set title
  theme_bw()+ #to get grey title box
  theme(axis.title.x=element_blank(),axis.text.x=element_blank(),axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),axis.text.y=element_blank(),axis.ticks.y=element_blank(),
        line = element_blank()) #to remove axes and axes lines
  

#lower in ctr
#sensitive bacteria gly
exp1_ASV_lower <- bchanged_gly_exp1$ASV_ID[bchanged_gly_exp1$`logFC ctr:hc`<0]
exp2_ASV_lower <- bchanged_gly_exp2$ASV_ID[bchanged_gly_exp2$`logFC ctr:hc`<0]
list_venn_gly <- list(exp1_lower = exp1_ASV_lower, exp2_lower = exp2_ASV_lower)
#relative abundance
bCRA_lower_gly_exp1 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","gly"), c(2,5), 1, bchanged_gly_exp1[bchanged_gly_exp1$`logFC ctr:hc` < 0,])[[1]]
bCRA_lower_gly_exp2 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","gly"), c(2,5), 2, bchanged_gly_exp2[bchanged_gly_exp2$`logFC ctr:hc` < 0,])[[1]]
#sensitive bacteria tb
exp1_ASV_lower <- bchanged_tb_exp1$ASV_ID[bchanged_tb_exp1$`logFC ctr:hc`<0]
exp2_ASV_lower <- bchanged_tb_exp2$ASV_ID[bchanged_tb_exp2$`logFC ctr:hc`<0]
list_venn_tb <- list(exp1_lower = exp1_ASV_lower, exp2_lower = exp2_ASV_lower)
#relative abundance
bCRA_lower_tb_exp1 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","tb"), c(2,5), 1, bchanged_tb_exp1[bchanged_tb_exp1$`logFC ctr:hc` < 0,])[[1]]
bCRA_lower_tb_exp2 <- cRA_Application(bDESIGN, bDAT_rare, c("ctr","tb"), c(2,5), 2, bchanged_tb_exp2[bchanged_tb_exp2$`logFC ctr:hc` < 0,])[[1]]

#labels
list_venn <- c(list_venn_gly, list_venn_tb)
labels <- c(paste0("exp 1\ncRA: ", round(bCRA_lower_gly_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(bCRA_lower_gly_exp2, 2), "%"),
            paste0("exp 1\ncRA: ", round(bCRA_lower_tb_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(bCRA_lower_tb_exp2, 2), "%"))
names(list_venn) <- labels

#plot
bvenn_l <- ggvenn(list_venn, fill_color = c(col_gly_up, col_gly_up, col_tb_up, col_tb_up), stroke_size = 0.5, set_name_size = 2, show_percentage = F) +
  facet_wrap(~"higher in hc")+ #to set title
  theme_bw()+ #to get grey title box
  theme(axis.title.x=element_blank(),axis.text.x=element_blank(),axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),axis.text.y=element_blank(),axis.ticks.y=element_blank(),
        line = element_blank()) #to remove axes and axes lines



#combine plots
fig7.1.2 <- plot_grid(
                 #NULL,
                 bvenn_h + theme(legend.position="none", plot.margin = unit(c(0, 0.1, 0, 0.1), "cm")),
                 bvenn_l+ theme(legend.position="none", plot.margin = unit(c(0, 0.2, 0, 0.1), "cm")) + labs(colour=""),
                 align = "hv",
                 #rel_widths = c(0.2,1,1),
                 nrow=1)

fig7.1 <- plot_grid(fig7.1.1 + theme(legend.position="none"),
                 fig7.1.2+ theme(legend.position="none") + labs(colour=""),
                 #labels=c("A","B"),
                 align = "hv",
                 rel_heights = c(1,.5),
                 nrow=2)

#saveRDS(fig7.1, "paper_figs/RDS/fig7.1.RDS")

#pseudoplot for legend
df <- data.frame(x=1:4, y=1:4, Herbicide=c("higher in ctr", "higher in gly", "higher in tb", "NS"))
pseudo <- ggplot(df, aes(x,y, col=Herbicide))+
            geom_point()+
            scale_color_manual(values=c(as.character(level_cols_Herbicide), "grey"))+
            theme_bw()

#add legend
legend <- get_legend(pseudo + theme(legend.position="bottom") + labs(col="Application"))
fig7.1 <- plot_grid(fig7.1, legend, nrow=2, rel_heights=c(1, .1))

#print
fig7.1

```

\pagebreak

### Fungi

```{r fun exp1 herbicide, warning=F, echo=F}

#exp1

#gly
temp <-   sensitive_to_herbicide(fDESIGN, fDAT, fTAX, c("ctr","gly"), c(2,5), 1, species = "fungi")
fEDGER_summary_gly_exp1 <- temp[[1]]
fchanged_gly_exp1 <- temp[[2]]
fEDGER_fit_lrt_table_ASVs_gly_exp1 <- temp[[3]]
fEDGER_fit_lrt_topTags_gly_exp1 <- temp[[4]]
lrt_fASVs_gly_exp1 <- temp[[5]]
fEDGER_fit_lrt_table_gly_exp1 <- temp[[6]]

temp <- cRA_Herbicide(fDESIGN, fDAT_rare, c("ctr","gly"), c(2,5), 1, fchanged_gly_exp1)
fCRA_gly_exp1 <- temp[[1]]
fchanged_gly_exp1 <- temp[[3]]

pander(fEDGER_summary_gly_exp1[,c(3,2,1)], justify='left', caption = "Fungi: Glyphosate effect in experiment 1") 
print(paste0(round(fCRA_gly_exp1, 2), "%", " of the fungal abundance changed after applying glyphosate"))
fchanged_gly_exp1_out <- fchanged_gly_exp1[, c(2,3,11,13,14)]
kbl(fchanged_gly_exp1_out[1:min(10, nrow(fchanged_gly_exp1_out)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")

#write.csv(fchanged_gly_exp1_out, "paper_tables/fchanged_gly_exp1.csv")

#tb
temp <-   sensitive_to_herbicide(fDESIGN, fDAT, fTAX, c("ctr","tb"), c(2,5), 1, species = "fungi")
fEDGER_summary_tb_exp1 <- temp[[1]]
fchanged_tb_exp1 <- temp[[2]]
fEDGER_fit_lrt_table_ASVs_tb_exp1 <- temp[[3]]
fEDGER_fit_lrt_topTags_tb_exp1 <- temp[[4]]
lrt_fASVs_tb_exp1 <- temp[[5]]
fEDGER_fit_lrt_table_tb_exp1 <- temp[[6]]

temp <- cRA_Herbicide(fDESIGN, fDAT_rare, c("ctr","tb"), c(2,5), 1, fchanged_tb_exp1)
fCRA_tb_exp1 <- temp[[1]]
fchanged_tb_exp1 <- temp[[3]]

pander(fEDGER_summary_tb_exp1[,c(3,2,1)], justify='left', caption = "Fungi: Terbuthylazine effect in experiment 1") 
print(paste0(round(fCRA_tb_exp1, 2), "%", " of the fungal abundance changed after applying terbuthylazine"))
fchanged_tb_exp1_out <- fchanged_tb_exp1[, c(2,3,11,13,14)]
kbl(fchanged_tb_exp1_out[1:min(10, nrow(fchanged_tb_exp1_out)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")

#write.csv(fchanged_tb_exp1_out, "paper_tables/fchanged_tb_exp1.csv")

```

\pagebreak

```{r fun exp2 herbicide, warning=F, echo=F}

#exp2

#gly
temp <-   sensitive_to_herbicide(fDESIGN, fDAT, fTAX, c("ctr","gly"), c(2,5), 2, species = "fungi")
fEDGER_summary_gly_exp2 <- temp[[1]]
fchanged_gly_exp2 <- temp[[2]]
fEDGER_fit_lrt_table_ASVs_gly_exp2 <- temp[[3]]
fEDGER_fit_lrt_topTags_gly_exp2 <- temp[[4]]
lrt_fASVs_gly_exp2 <- temp[[5]]
fEDGER_fit_lrt_table_gly_exp2 <- temp[[6]]

temp <- cRA_Herbicide(fDESIGN, fDAT_rare, c("ctr","gly"), c(2,5), 2, fchanged_gly_exp2)
fCRA_gly_exp2 <- temp[[1]]
fchanged_gly_exp2 <- temp[[3]]

pander(fEDGER_summary_gly_exp2[,c(3,2,1)], justify='left', caption = "Fungi: Glyphosate effect in experiment 2") 
print(paste0(round(fCRA_gly_exp2, 2), "%", " of the fungal abundance changed after applying glyphosate"))
fchanged_gly_exp2_out <- fchanged_gly_exp2[, c(2,3,11,13,14)]
kbl(fchanged_gly_exp2_out[1:min(10, nrow(fchanged_gly_exp2_out)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")

#write.csv(fchanged_gly_exp2_out, "paper_tables/fchanged_gly_exp2.csv")

#tb
temp <-   sensitive_to_herbicide(fDESIGN, fDAT, fTAX, c("ctr","tb"), c(2,5), 2, species = "fungi")
fEDGER_summary_tb_exp2 <- temp[[1]]
fchanged_tb_exp2 <- temp[[2]]
fEDGER_fit_lrt_table_ASVs_tb_exp2 <- temp[[3]]
fEDGER_fit_lrt_topTags_tb_exp2 <- temp[[4]]
lrt_fASVs_tb_exp2 <- temp[[5]]
fEDGER_fit_lrt_table_tb_exp2 <- temp[[6]]

temp <- cRA_Herbicide(fDESIGN, fDAT_rare, c("ctr","tb"), c(2,5), 2, fchanged_tb_exp2)
fCRA_tb_exp2 <- temp[[1]]
fchanged_tb_exp2 <- temp[[3]]

pander(fEDGER_summary_tb_exp2[,c(3,2,1)], justify='left', caption = "Fungi: Terbuthylazine effect in experiment 2") 
print(paste0(round(fCRA_tb_exp2, 2), "%", " of the fungal abundance changed after applying terbuthylazine"))
fchanged_tb_exp2_out <- fchanged_tb_exp2[, c(2,3,11,13,14)]
kbl(fchanged_tb_exp2_out[1:min(10, nrow(fchanged_tb_exp2_out)),], booktabs = T, align = "c", linesep = "", caption="The top abundant ASVs")%>%
  kable_styling(latex_options = "HOLD_position")

#write.csv(fchanged_tb_exp2_out, "paper_tables/fchanged_tb_exp2.csv")

```

Glyphosate changed the abundance of `r paste(length(fEDGER_fit_lrt_table_ASVs_gly_exp1))` bacterial ASVs in experiment 1 and `r paste(length(fEDGER_fit_lrt_table_ASVs_gly_exp2))` bacterial ASVs in experiment 2. Applying terbuthylazine led to changes in abundance for `r paste(length(fEDGER_fit_lrt_table_ASVs_tb_exp1))` bacterial ASVs in experiment 1 and `r paste(length(fEDGER_fit_lrt_table_ASVs_tb_exp2))` bacterial ASVs in experiment 2.

### Figure 7.2 | Fungi: MA & venn

\vspace{5mm}

```{r MA and venn fungi herbicide, echo=F, warning=F, message=F, eval=T, fig.height=10, fig.width=6}

#gly
#create fungi df
df1 <- fEDGER_fit_lrt_topTags_gly_exp1[["table"]]
df1$herbicide <- "gly"
df1$experiment <- "Experiment 1"
df2 <- fEDGER_fit_lrt_topTags_gly_exp2[["table"]]
df2$herbicide <- "gly"
df2$experiment <- "Experiment 2"

df <- rbind(df1, df2)
rm(df1, df2)

#get up and down-regulated ASVs
df$cols <- "NS"
#df$logFC <- df$logFC * (-1) #change from hc:ctr to ctr:hc #change from hc:ctr to ctr:hc
df$cols[df$FDR < 0.05 & df$logFC > 0] <- "higher in ctr"
df$cols[df$FDR < 0.05 & df$logFC < 0] <- "higher in gly"

#sort to plot sign at the end
df <- df[order(df$FDR, decreasing = T),]

#plot
fig7.2.1_gly <- ggplot(df, aes(x=logCPM, y=logFC))+
                  geom_point(aes(col=cols), alpha=0.8, size=2)+
                  scale_color_manual(values=c(col_ctr_up,col_gly_up, "grey"))+
                  facet_wrap(~ experiment)+
                  theme_bw()+
                  ylab("log2 fold change")+
                  xlab("log2 mean abundance")

#tb
#create fungi df
df1 <- fEDGER_fit_lrt_topTags_tb_exp1[["table"]]
df1$herbicide <- "tb"
df1$experiment <- "Experiment 1"
df2 <- fEDGER_fit_lrt_topTags_tb_exp2[["table"]]
df2$herbicide <- "tb"
df2$experiment <- "Experiment 2"

df <- rbind(df1, df2)
rm(df1, df2)

#get up and down-regulated ASVs
df$cols <- "NS"
#df$logFC <- df$logFC * (-1) #change from hc:ctr to ctr:hc #change from hc:ctr to ctr:hc
df$cols[df$FDR < 0.05 & df$logFC > 0] <- "higher in ctr"
df$cols[df$FDR < 0.05 & df$logFC < 0] <- "higher in tb"

#sort to plot sign at the end
df <- df[order(df$FDR, decreasing = T),]

#plot
fig7.2.1_tb <- ggplot(df, aes(x=logCPM, y=logFC))+
                  geom_point(aes(col=cols), alpha=0.8, size=2)+
                  scale_color_manual(values=c(col_ctr_up,col_tb_up, "grey"))+
                  facet_wrap(~ experiment)+
                  theme_bw()+
                  ylab("log2 fold change")+
                  xlab("log2 mean abundance")

#combine MA plots
fig7.2.1 <- plot_grid(fig7.2.1_gly + theme(legend.position="none", plot.margin = unit(c(0.2, 0.2, 0.2, 0.2), "cm")),
                 fig7.2.1_tb+ theme(legend.position="none", plot.margin = unit(c(0.2, 0.2, 0.2, 0.2), "cm")) + labs(colour=""),
                 #labels=c("A","B"),
                 align = "hv",
                 nrow=2)

#VENN

#higher in ctr
#sensitive fungi gly
exp1_ASV_higher <- fchanged_gly_exp1$ASV_ID[fchanged_gly_exp1$`logFC ctr:hc`>0]
exp2_ASV_higher <- fchanged_gly_exp2$ASV_ID[fchanged_gly_exp2$`logFC ctr:hc`>0]
list_venn_gly <- list(exp1_higher = exp1_ASV_higher, exp2_higher = exp2_ASV_higher)
#relative abundance
fCRA_higher_gly_exp1 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","gly"), c(2,5), 1, fchanged_gly_exp1[fchanged_gly_exp1$`logFC ctr:hc` > 0,])[[1]]
fCRA_higher_gly_exp2 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","gly"), c(2,5), 2, fchanged_gly_exp2[fchanged_gly_exp2$`logFC ctr:hc` > 0,])[[1]]
#sensitive fungi tb
exp1_ASV_higher <- fchanged_tb_exp1$ASV_ID[fchanged_tb_exp1$`logFC ctr:hc`>0]
exp2_ASV_higher <- fchanged_tb_exp2$ASV_ID[fchanged_tb_exp2$`logFC ctr:hc`>0]
list_venn_tb <- list(exp1_higher = exp1_ASV_higher, exp2_higher = exp2_ASV_higher)
#relative abundance
fCRA_higher_tb_exp1 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","tb"), c(2,5), 1, fchanged_tb_exp1[fchanged_tb_exp1$`logFC ctr:hc` > 0,])[[1]]
fCRA_higher_tb_exp2 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","tb"), c(2,5), 2, fchanged_tb_exp2[fchanged_tb_exp2$`logFC ctr:hc` > 0,])[[1]]

#labels
list_venn <- c(list_venn_gly, list_venn_tb)
labels <- c(paste0("exp 1\ncRA: ", round(fCRA_higher_gly_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(fCRA_higher_gly_exp2, 2), "%"),
            paste0("exp 1\ncRA: ", round(fCRA_higher_tb_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(fCRA_higher_tb_exp2, 2), "%"))
names(list_venn) <- labels

#plot
fvenn_h <- ggvenn(list_venn, fill_color = c(col_gly_up, col_gly_up, col_tb_up, col_tb_up), stroke_size = 0.5, set_name_size = 2, show_percentage = F) +
  facet_wrap(~"higher in ctr")+ #to set title
  theme_bw()+ #to get grey title box
  theme(axis.title.x=element_blank(),axis.text.x=element_blank(),axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),axis.text.y=element_blank(),axis.ticks.y=element_blank(),
        line = element_blank()) #to remove axes and axes lines
  

#lower in ctr
#sensitive fungi gly
exp1_ASV_lower <- fchanged_gly_exp1$ASV_ID[fchanged_gly_exp1$`logFC ctr:hc`<0]
exp2_ASV_lower <- fchanged_gly_exp2$ASV_ID[fchanged_gly_exp2$`logFC ctr:hc`<0]
list_venn_gly <- list(exp1_lower = exp1_ASV_lower, exp2_lower = exp2_ASV_lower)
#relative abundance
fCRA_lower_gly_exp1 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","gly"), c(2,5), 1, fchanged_gly_exp1[fchanged_gly_exp1$`logFC ctr:hc` < 0,])[[1]]
fCRA_lower_gly_exp2 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","gly"), c(2,5), 2, fchanged_gly_exp2[fchanged_gly_exp2$`logFC ctr:hc` < 0,])[[1]]
#sensitive fungi tb
exp1_ASV_lower <- fchanged_tb_exp1$ASV_ID[fchanged_tb_exp1$`logFC ctr:hc`<0]
exp2_ASV_lower <- fchanged_tb_exp2$ASV_ID[fchanged_tb_exp2$`logFC ctr:hc`<0]
list_venn_tb <- list(exp1_lower = exp1_ASV_lower, exp2_lower = exp2_ASV_lower)
#relative abundance
fCRA_lower_tb_exp1 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","tb"), c(2,5), 1, fchanged_tb_exp1[fchanged_tb_exp1$`logFC ctr:hc` < 0,])[[1]]
fCRA_lower_tb_exp2 <- cRA_Application(fDESIGN, fDAT_rare, c("ctr","tb"), c(2,5), 2, fchanged_tb_exp2[fchanged_tb_exp2$`logFC ctr:hc` < 0,])[[1]]

#labels
list_venn <- c(list_venn_gly, list_venn_tb)
labels <- c(paste0("exp 1\ncRA: ", round(fCRA_lower_gly_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(fCRA_lower_gly_exp2, 2), "%"),
            paste0("exp 1\ncRA: ", round(fCRA_lower_tb_exp1, 2), "%"), 
            paste0("exp 2\ncRA: ", round(fCRA_lower_tb_exp2, 2), "%"))
names(list_venn) <- labels

#plot
fvenn_l <- ggvenn(list_venn, fill_color = c(col_gly_up, col_gly_up, col_tb_up, col_tb_up), stroke_size = 0.5, set_name_size = 2, show_percentage = F) +
  facet_wrap(~"higher in hc")+ #to set title
  theme_bw()+ #to get grey title box
  theme(axis.title.x=element_blank(),axis.text.x=element_blank(),axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),axis.text.y=element_blank(),axis.ticks.y=element_blank(),
        line = element_blank()) #to remove axes and axes lines



#combine plots
fig7.2.2 <- plot_grid(
                 #NULL,
                 fvenn_h + theme(legend.position="none", plot.margin = unit(c(0, 0.1, 0, 0.1), "cm")),
                 fvenn_l+ theme(legend.position="none", plot.margin = unit(c(0, 0.2, 0, 0.1), "cm")) + labs(colour=""),
                 #rel_widths = c(0.2,1,1),
                 align = "hv",
                 nrow=1)

fig7.2 <- plot_grid(fig7.2.1 + theme(legend.position="none"),
                 fig7.2.2+ theme(legend.position="none") + labs(colour=""),
                 #labels=c("A","B"),
                 align = "hv",
                 rel_heights = c(1,.5),
                 nrow=2)

#saveRDS(fig7.2, "paper_figs/RDS/fig7.2.RDS")

#add legend
legend <- get_legend(pseudo + theme(legend.position="bottom") + labs(col="Application"))
fig7.2 <- plot_grid(fig7.2, legend, nrow=2, rel_heights=c(1, .1))

#print
fig7.2

```

```{r abu, echo=F, warning=F, message=F, eval=T}

# #check abu for paper
# 
# #check ASVs sensitive for gly and tb
# all_exp1 <- c(rownames(bchanged_gly_exp1), rownames(bchanged_tb_exp1))
# common_exp1 <- all_exp1[duplicated(all_exp1)]
# 
# abu_exp1 <- bchanged_gly_exp1[bchanged_gly_exp1$ASV_ID %in% common_exp1, c("ASV_ID", "rAbu")]
# tax_exp1 <- bTAX[bTAX$ASV_ID %in% common_exp1,]
# #no high abundant
# 
# 
# all_exp2 <- c(rownames(bchanged_gly_exp2), rownames(bchanged_tb_exp2))
# common_exp2 <- all_exp2[duplicated(all_exp2)]
# 
# abu_exp2 <- bchanged_gly_exp2[bchanged_gly_exp2$ASV_ID %in% common_exp2, c("ASV_ID", "rAbu")]
# tax_exp2 <- bTAX[bTAX$ASV_ID %in% common_exp2,]

```


\vspace{5mm}

**Conclusion:** Applying glyphosate and terbuthylazine led to similar effect. In both experiments were around 30 ASVs sensitive in their abundance to glyphosate and terbuthylazine but non of those ASVs was sensitive in both experiments. We found high abundant bacterial ASVs reacting sensitive to herbicides which was up to 12 % of the bacterial community. Only a few fungal ASVs were changed in their abundance and non of them was sensitive for both herbicides or in both experiments.

\vspace{5mm}

## Conclusion taxa response

Applying herbicides to different targets did not influence high abundant ASVs. Applying glyphosate and terbuthylazine caused similar effects. Both altered, multiple bacterial ASVs. Almost no effect was found on fungi. Sensitive ASVs were not consistent between the two experiments.
